ai-service/requirements.txt
# AI Surveillance Co-Pilot - Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3

# YOLO Object Detection
ultralytics==8.1.0
torch==2.1.0
torchvision==0.16.0

# HTTP Client
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Logging
colorlog==6.8.0

# Optional: For GPU support (uncomment if using CUDA)
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118
# -f https://download.pytorch.org/whl/torch_stable.html